chore: Add playground project
