--- examples/basic/main.ts
@@ -1,5 +1,0 @@
-import { Xeito } from './../../packages/core';
-import { AppComponent } from './app-component';
-
-const app = new Xeito(AppComponent);
-app.bootstrap('#app');
