+++ playground/xeito-base/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "xeito-base",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0"
+  },
+  "dependencies": {
+    "@xeito/vite-plugin-xeito": "^0.0.3",
+    "vite-plugin-inspect": "^0.7.15"
+  }
+}
