+++ playground/xeito-base/src/app.component.ts
@@ -1,0 +1,17 @@
+import { XeitoComponent, Component, State, html } from "../../../packages/core";
+
+@Component({
+  selector: 'app-root'
+})
+export class AppComponent extends XeitoComponent {
+
+  @State() count: number = 0;
+
+  render() {
+    return html`
+      <h1>Xeito default playground</h1>
+      <button @click=${()=>this.count++}>Count is: ${this.count}</button>
+    `;
+  }
+
+}
