+++ playground/xeito-base/src/main.ts
@@ -1,0 +1,5 @@
+import { AppComponent } from './app.component';
+import { Xeito } from '@xeito/core';
+
+const app = new Xeito(AppComponent);
+app.bootstrap('#app');
